+++ uoxserver/UOXJSPropertyFuncs.cpp
@@ -342,7 +342,7 @@ JSBool CCreateEntryProps_getProperty( JSContext *cx, JSObject *obj, jsval id, js
 			jsval amountNeeded = 0;
 			jsval targColour = 0;
 			JSObject *resources = JS_NewArrayObject( cx, 0, nullptr );
-			for( int i = 0; i < resourcesNeeded.size(); i++ )
+			for( size_t i = 0; i < resourcesNeeded.size(); i++ )
 			{
 				resEntry = resourcesNeeded[i];
 				amountNeeded = INT_TO_JSVAL( resEntry.amountNeeded );
@@ -355,7 +355,7 @@ JSBool CCreateEntryProps_getProperty( JSContext *cx, JSObject *obj, jsval id, js
 
 				// Make a child object with all the supported resource IDs
 				JSObject *idList = JS_NewArrayObject( cx, 0, nullptr );
-				for( int j = 0; j < resEntry.idList.size(); j++ )
+				for( size_t j = 0; j < resEntry.idList.size(); j++ )
 				{
 					jsval targID = INT_TO_JSVAL( resEntry.idList[j] );
 					JS_SetElement( cx, idList, j, &targID );
@@ -381,7 +381,7 @@ JSBool CCreateEntryProps_getProperty( JSContext *cx, JSObject *obj, jsval id, js
 			jsval minSkill = 0;
 			jsval maxSkill = 0;
 			JSObject *skills = JS_NewArrayObject( cx, 0, nullptr );
-			for( int i = 0; i < skillReqs.size(); i++ )
+			for( size_t i = 0; i < skillReqs.size(); i++ )
 			{
 				resEntry = skillReqs[i];
 				skillNumber = INT_TO_JSVAL( resEntry.skillNumber );
@@ -500,7 +500,7 @@ JSBool CItemProps_getProperty( JSContext *cx, JSObject *obj, jsval id, jsval *vp
 				JSObject *scriptTriggersJS = JS_NewArrayObject(cx, 0, nullptr);
 
 				std::vector<UI16> scriptTriggers = gPriv->GetScriptTriggers();
-				for( int i = 0; i < scriptTriggers.size(); i++ )
+				for( size_t i = 0; i < scriptTriggers.size(); i++ )
 				{
 					scriptID = INT_TO_JSVAL( scriptTriggers[i] );
 					JS_SetElement(cx, scriptTriggersJS, i, &scriptID);
@@ -1357,7 +1357,7 @@ JSBool CCharacterProps_getProperty( JSContext *cx, JSObject *obj, jsval id, jsva
 				JSObject *scriptTriggersJS = JS_NewArrayObject(cx, 0, nullptr);
 
 				std::vector<UI16> scriptTriggers = gPriv->GetScriptTriggers();
-				for( int i = 0; i < scriptTriggers.size(); i++ )
+				for( size_t i = 0; i < scriptTriggers.size(); i++ )
 				{
 					scriptID = INT_TO_JSVAL( scriptTriggers[i] );
 					JS_SetElement(cx, scriptTriggersJS, i, &scriptID);
@@ -2217,7 +2217,7 @@ JSBool CRegionProps_getProperty( JSContext *cx, JSObject *obj, jsval id, jsval *
 				JSObject *scriptTriggersJS = JS_NewArrayObject(cx, 0, nullptr);
 
 				std::vector<UI16> scriptTriggers = gPriv->GetScriptTriggers();
-				for( int i = 0; i < scriptTriggers.size(); i++ )
+				for( size_t i = 0; i < scriptTriggers.size(); i++ )
 				{
 					scriptID = INT_TO_JSVAL( scriptTriggers[i] );
 					JS_SetElement(cx, scriptTriggersJS, i, &scriptID);
