+++ uoxserver/items.cpp
@@ -754,7 +754,7 @@ CItem *cItem::CreateRandomItem( CItem *mCont, const std::string& sItemList, cons
 			//	weight|LOOTLIST=sectionName,amountMin amountMax
 			//	weight|ITEMLIST=sectionName,amount
 			//	weight|ITEMLIST=sectionName,amountMin amountMax
-			for( int j = 0; j < itemListSize; j++ )
+			for( int j = 0; j < static_cast<int>(itemListSize); j++ )
 			{
 				// Split string for entry into a stringlist, based on | as a separator
 				auto csecs = oldstrutil::sections( oldstrutil::trim( oldstrutil::removeTrailing( ItemList->MoveTo( j ), "//" )), "|" );
@@ -779,7 +779,7 @@ CItem *cItem::CreateRandomItem( CItem *mCont, const std::string& sItemList, cons
 
 			// Loop through the items in the itemlist/lootlist
 			int weightOfChosenItem = 0;
-			for( int j = 0; j < itemListSize; j++ )
+			for( int j = 0; j < static_cast<int>(itemListSize); j++ )
 			{
 				auto csecs = oldstrutil::sections( oldstrutil::trim( oldstrutil::removeTrailing( ItemList->MoveTo( j ), "//" )), "|" );
 				if( csecs.size() == 2 )
